+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/CategorySearchTest.kt
@@ -216,68 +216,33 @@ internal class CategorySearchTest {
     }
 
     @TestFactory
-    fun `Check consecutive search calls`() = TestCase {
-        Given("GeocodingSearchEngine with mocked dependencies") {
+    fun `Check search call cancellation`() = TestCase {
+        Given("SearchEngine with mocked dependencies") {
             every { searchResultFactory.createSearchResult(any(), any()) } returns TEST_SEARCH_RESULT
 
-            val options = CategorySearchOptions()
+            val cancellationReason = "Request cancelled"
 
-            val category1 = "cafe"
-            val slotSearchCallback1 = slot<CoreSearchCallback>()
-            every { coreEngine.search(eq(""), eq(listOf(category1)), any(), capture(slotSearchCallback1)) } returns Unit
-
-            val category2 = "museum"
-            val slotSearchCallback2 = slot<CoreSearchCallback>()
-            every { coreEngine.search(eq(""), eq(listOf(category2)), any(), capture(slotSearchCallback2)) } returns Unit
-
-            var searchRequestTask1: SearchRequestTaskImpl<*>? = null
-            When("Search function called for the first time") {
-                val callback1 = mockk<SearchCallback>(relaxed = true)
-                searchRequestTask1 = (searchEngine.search(category1, options, callback1) as? SearchRequestTaskImpl<*>)
-
-                Then("First task is not executed", false, searchRequestTask1?.isDone)
-                Then("First task is not cancelled", false, searchRequestTask1?.isCancelled)
-                Then(
-                    "First task keeps original listener",
-                    true,
-                    searchRequestTask1?.callbackDelegate != null
-                )
+            val slotSearchCallback = slot<CoreSearchCallback>()
+            every { coreEngine.search(eq(""), eq(listOf(TEST_CATEGORIES_QUERY)), any(), capture(slotSearchCallback)) } answers {
+                slotSearchCallback.captured.run(createTestCoreSearchResponseCancelled(cancellationReason))
             }
 
-            var searchRequestTask2: SearchRequestTaskImpl<*>? = null
-            When("Search function called for the second time and first request automatically cancelled") {
-                slotSearchCallback1.captured.run(createTestCoreSearchResponseCancelled())
+            When("Search request cancelled by the Search SDK") {
+                val callback = mockk<SearchCallback>(relaxed = true)
 
-                val callback2 = mockk<SearchCallback>(relaxed = true)
-                searchRequestTask2 = (searchEngine.search(category2, options, callback2) as? SearchRequestTaskImpl<*>)
+                val task = (searchEngine.search(TEST_CATEGORIES_QUERY, TEST_SEARCH_OPTIONS, callback) as? SearchRequestTaskImpl<*>)
 
-                Then("First task is not executed", false, searchRequestTask1?.isDone)
-                Then("First task is cancelled", true, searchRequestTask1?.isCancelled)
+                Then("Task is not executed", false, task?.isDone)
+                Then("Task is cancelled", true, task?.isCancelled)
                 Then(
-                    "First task still keeps reference to original listener",
-                    false,
-                    searchRequestTask1?.callbackDelegate != null
-                )
-
-                Then("Second task is not executed", false, searchRequestTask2?.isDone)
-                Then("Second task is not cancelled", false, searchRequestTask2?.isCancelled)
-                Then(
-                    "Second task keeps original listener",
+                    "Task released reference to original callback",
                     true,
-                    searchRequestTask2?.callbackDelegate != null
+                    task?.callbackDelegate == null
                 )
-            }
 
-            When("Second search request completes") {
-                slotSearchCallback2.captured.run(TEST_SUCCESSFUL_CORE_RESPONSE)
-
-                Then("Second task is executed", true, searchRequestTask2?.isDone)
-                Then("Second task is not cancelled", false, searchRequestTask2?.isCancelled)
-                Then(
-                    "Second task released reference to original listener",
-                    true,
-                    searchRequestTask2?.callbackDelegate == null
-                )
+                VerifyOnce("Callback called with cancellation error") {
+                    callback.onError(eq(SearchCancellationException(cancellationReason)))
+                }
             }
         }
     }