+++ MapboxSearch/sdk/src/test/java/com/mapbox/search/SearchEngineTest.kt
@@ -232,68 +232,31 @@ internal class SearchEngineTest {
     }
 
     @TestFactory
-    fun `Check consecutive search calls`() = TestCase {
+    fun `Check search call cancellation`() = TestCase {
         Given("SearchEngine with mocked dependencies") {
-            val searchQuery1 = "Query 1"
-            val slotSearchCallback1 = slot<CoreSearchCallback>()
-            every { coreEngine.search(eq(searchQuery1), any(), any(), capture(slotSearchCallback1)) } returns Unit
+            val cancellationReason = "Request cancelled"
 
-            val searchQuery2 = "Query 2"
-            val slotSearchCallback2 = slot<CoreSearchCallback>()
-            every { coreEngine.search(eq(searchQuery2), any(), any(), capture(slotSearchCallback2)) } returns Unit
-
-            var searchRequestTask1: SearchRequestTaskImpl<*>? = null
-            When("Search function called for the first time") {
-                val callback1 = mockk<SearchSuggestionsCallback>(relaxed = true)
-
-                searchRequestTask1 =
-                    (searchEngine.search(searchQuery1, TEST_SEARCH_OPTIONS, callback1) as? SearchRequestTaskImpl<*>)
-
-                Then("First task is not executed", false, searchRequestTask1?.isDone)
-                Then("First task is not cancelled", false, searchRequestTask1?.isCancelled)
-                Then(
-                    "First task keeps original callback",
-                    true,
-                    searchRequestTask1?.callbackDelegate != null
-                )
+            val slotSearchCallback = slot<CoreSearchCallback>()
+            every { coreEngine.search(eq(TEST_QUERY), any(), any(), capture(slotSearchCallback)) } answers {
+                slotSearchCallback.captured.run(createTestCoreSearchResponseCancelled(cancellationReason))
             }
 
-            var searchRequestTask2: SearchRequestTaskImpl<*>? = null
-            When("Search function called for the second time and first request automatically cancelled") {
-                slotSearchCallback1.captured.run(createTestCoreSearchResponseCancelled())
+            When("Search request cancelled by the Search SDK") {
+                val callback = mockk<SearchSuggestionsCallback>(relaxed = true)
 
-                val callback2 = mockk<SearchSuggestionsCallback>(relaxed = true)
+                val task = (searchEngine.search(TEST_QUERY, TEST_SEARCH_OPTIONS, callback) as? SearchRequestTaskImpl<*>)
 
-                searchRequestTask2 =
-                    (searchEngine.search(searchQuery2, TEST_SEARCH_OPTIONS, callback2) as? SearchRequestTaskImpl<*>)
-
-                Then("First task is not executed", false, searchRequestTask1?.isDone)
-                Then("First task is cancelled", true, searchRequestTask1?.isCancelled)
+                Then("Task is not executed", false, task?.isDone)
+                Then("Task is cancelled", true, task?.isCancelled)
                 Then(
-                    "First task released reference to original callback",
+                    "Task released reference to original callback",
                     true,
-                    searchRequestTask1?.callbackDelegate == null
+                    task?.callbackDelegate == null
                 )
 
-                Then("Second task is not executed", false, searchRequestTask2?.isDone)
-                Then("Second task is not cancelled", false, searchRequestTask2?.isCancelled)
-                Then(
-                    "Second task keeps original callback",
-                    true,
-                    searchRequestTask2?.callbackDelegate != null
-                )
-            }
-
-            When("Second search request completes") {
-                slotSearchCallback2.captured.run(TEST_SUCCESSFUL_CORE_RESPONSE)
-
-                Then("Second task is executed", true, searchRequestTask2?.isDone)
-                Then("Second task is not cancelled", false, searchRequestTask2?.isCancelled)
-                Then(
-                    "Second task released reference to original callback",
-                    true,
-                    searchRequestTask2?.callbackDelegate == null
-                )
+                VerifyOnce("Callback called with cancellation error") {
+                    callback.onError(eq(SearchCancellationException(cancellationReason)))
+                }
             }
         }
     }